get all lotteries test
