+++ posttest/build.gradle
@@ -34,6 +34,12 @@ dependencies {
 	testImplementation 'org.springframework.security:spring-security-test'
 
 	implementation 'org.springframework.boot:spring-boot-starter-validation'
+
+	testImplementation platform('org.junit:junit-bom:5.9.1')
+	testImplementation 'org.junit.jupiter:junit-jupiter'
+
+	testImplementation 'org.mockito:mockito-core:5.6.0'
+	implementation "io.reactivex.rxjava3:rxjava:3.1.8"
 }
 
 tasks.named('test') {
