+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/lottery/Lottery.java
@@ -18,7 +18,7 @@ public class Lottery {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @Column(name = "ticket")
     private String ticket;
