+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/lottery/LotteryController.java
@@ -1,11 +1,13 @@
 package com.kbtg.bootcamp.posttest.lottery;
 
 import com.kbtg.bootcamp.posttest.exception.InternalServiceException;
+import com.kbtg.bootcamp.posttest.lottery.dto.LotteryListResponseDto;
 import com.kbtg.bootcamp.posttest.lottery.dto.LotteryRequestDto;
 import com.kbtg.bootcamp.posttest.lottery.dto.LotteryResponseDto;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
@@ -17,6 +19,16 @@ public class LotteryController {
 
     public LotteryController(LotteryService adminService) {
         this.lotteryService = adminService;
+    }
+
+    @GetMapping("/lotteries")
+    public ResponseEntity<LotteryListResponseDto> getLotteryList() {
+        try {
+            LotteryListResponseDto response = lotteryService.getAllLotteries();
+            return ResponseEntity.ok(response);
+        } catch (Exception exception) {
+            throw new InternalServiceException("Internal service exception with Normal service");
+        }
     }
 
     @PostMapping("/admin/lotteries")
