+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/lottery/LotteryService.java
@@ -1,8 +1,11 @@
 package com.kbtg.bootcamp.posttest.lottery;
 
+import com.kbtg.bootcamp.posttest.lottery.dto.LotteryListResponseDto;
 import com.kbtg.bootcamp.posttest.lottery.dto.LotteryRequestDto;
 import com.kbtg.bootcamp.posttest.lottery.dto.LotteryResponseDto;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 @Service
 public class LotteryService {
@@ -10,6 +13,15 @@ public class LotteryService {
     final LotteryRepository lotteryRepository;
     public LotteryService(LotteryRepository lotteryRepository) {
         this.lotteryRepository = lotteryRepository;
+    }
+
+    public LotteryListResponseDto getAllLotteries() {
+        List<String> tickets = lotteryRepository.findAll()
+                .stream()
+                .filter(lottery -> lottery.getAmount() >= 1)
+                .map(Lottery::getTicket)
+                .toList();
+        return new LotteryListResponseDto(tickets);
     }
 
     public LotteryResponseDto createLottery (LotteryRequestDto requestDto){
